Make selection and order of buttons and menus on the tool shelf run-time configurable.
[Fixes #69466388] [Fixes #73645766]

We add to the CODAP web page inclusion of a javascript page ../../../../codap-config.js.
If this js file is present in the root of the built CODAP directory, it will
be included in the page. It can be used to alter the DG.toolButtons array
from core.js to alter the selection and order of buttons on the tool shelf.
This will take effect at run-time on an instance by instance basis.