[Fixes 78067876] Change path element of extended CODAP URL path to a constant string.
It had been the particular build number.

The relocatable version of codap works this way: In the build
script, makeCodapZip we run sproutcore's build tool,
sc-build. Sc-build builds a static collection of web pages
that wants to live at the root of the web server and which has
internal absolute links. So, we hunt all these links down and
convert them to relative links.