[Fixes #74002794] Add parent and child collection case counts to title bar.
For demonstration:
http://codap.concord.org/~jsandoe/sept10/static/dg/en/cert/index.html?moreGames=[{%22name%22:%22terapop%22,%22dimensions%22:{%22width%22:550,%22height%22:500},%20%22url%22:%22http://app.codap.concord.org/interactives/TerraPopInteractive/TPI.html%22}]

Handle setting counts in case tables initialized by documents.

We found that it is possible that the component is not
initialized when the case_table_controller initially attempts
to set the status. This happens when the table is initialized
during document initialization, such as in the Brazil
document.