[Fixes #78068692] Provide a one step deploy for developer builds.
The script is bin/deployDevCodap. It builds a portable version of the static
CODAP web app and uploads it to the user's public web directory on
the CODAP server, codap.concord.org. The directory is assumed to have been
set up. It is assumed to be public_html. After this is executed, the
version will be accessible at http://codap.concord.org/~{username}/{version-string}
where {version-string} is a name provided by the user. This link can be shared
with others to demonstrate prototypes or new features. If the command is
executed twice with the same version-string, the second time will overwrite
the first instance.