[Fixes 83356432] Child table creates new groups out of order.
This was because of new ID generation policy. Child table group order is in
order of parent IDs. Fixed by making sure new IDs are greater than existing
IDs.
The better fix would be to order child table groups by parent order, but
Case Tables are due to be replaced.