[Fixes #88623622] Revise logic of current game state to always call callback.
We were failing to call callback on errors.
Even if the DI fails to return a result, we want to continue the save.
Also, add polyfill for Promises, so logic will work in IE.
(In discussions with Aaron, we realized that there is already
usage of jQuery deferred objects in DocumentController. This is functionally equivalent to
promises. It doesn't make sense to use these two technologies in the same class.
Usages should be unified.)