Fix cause of broken build:
ES6 promises include a 'catch' method. Catch is a reserved word.
Sc-build uses yuicompressor which in turn uses, I think, the BISON compiler, which
was complaining about the use of the reserved word in an unexpected
place. This apparently threw an exception that aborted a portion of the sc-build
before some pngs were written out. The manifestation of this issue is transparent
button and menu backgrounds.