Fix: Clicking on point in graph or map generates error when CODAP attempts to log the case selected event.
Basically, it appears that when log manager responds with any error it is interpreted as a same origin violation.
I found that the log messages in question had a simple value for message argument. This is translated into the
value of the 'parameter' property which, on the Log Manager is expecting an object. I observed in the heroku
logs for the log manager that a non-object parameter value always produces an error.

Fixed by wrapping non-object parameters as a value.
[Fixes #90201376]