Fix incorrect appearance of user entry dialog when a document to load is already known.
This is the third attempt. We had looked for evidence of a document in the url.
On chromebooks, I assume because of timing, there is a point where the original doc id is removed, but the new doc id is not yet known. This always occurs in the LARA context when a runkey is present. So, we also look for runkey.
[#93849168]