Fix doRescaleAxesFromData to only animate change if requested
Previously, the function doRescaleAxesFromData always animated,
even if iAnimatePoints was false.

It looks like the previous implementation contained a typo:

```
if( iAnimatePoints)
        DG.sounds.playMixup();     // <- this is the only statement controlled by iAnimatePoints
        this.set('isAnimating', true); // <- this looks like it was intended to be as well

while( (iPlaces.length > 0) && (tOldBoundsArray.length > 0)) {  // <- and this is what animates
  // ... blah blah
```

Since the entire block is for the animation, the clearest fix seemed to
be to check iAnimatePoints in the top if-statement.

Now when iAnimatePoints is false, the axes rescale without animating.

Note that this will change the current behavior, because iAnimatePoints
*is* explicitly set to be false here:

https://github.com/concord-consortium/codap/blob/master/apps/dg/components/graph/plots/scatter_plot_model.js#L191