Extract game state saving from DocumentController to GameController
When a document closes, the document controller tries to save the state
of each DataInteractive.

This function is fairly involved, and depends on a lot of privileged expectations
of the GameController, so it makes more sense for it to be owned completely
by the GameController itself.

This has the benefit of creating a saveGameState function that can be called
on a single DataInteractive's controller, which can later be used to create
an undoable action when a DataInteractive is closed, allowing us to restore
that interactive with the saved game state (next commit).

(Note, this is a simple refactoring, so it retains the option of using the
apparently-obsolete `doCommandFunc`. But now that this is extracted to the
GameController, a later removal of doCommandFunc will be nicer, as it will
be completely localized to the GameController.)

[#96701234]