Undo/Redo: Support handling actions from external games (new API)
If we receive an `undoableActionPerformed` event, then we store a new
command in our undo history.

This command doesn't execute anything (because the game already performed
the action), but if we call undo/redo on the command, we call undo/redo
on the game, using the new commands `undoAction` and `redoAction`.

The API changes are as follows:

* Data Interactive-Initiated Commands
  * undoableActionPerformed

* CODAP-Initiated Actions
  * undoAction
  * redoAction

[#96197680]