Fix delete cases Redo.
This was tricky, because the original cases are gone for good after
they are deleted. What we want to "redo" is to delete the *new* cases
that were created by the undo.

We also need to keep up to date with each undo/redo cycle: if we undo
and redo multiple times, we need to keep track of which is the most
recent case that has been created that we need to delete.

[#96680770]