Bug Fix: Problems with slider undo.
Some of these problems had more to do with the existing slider implementation than with undo.
For example, close of the slider component was basically unimplemented. Important slider
actions did not dirty the document.

Undoing a slider open deletes the global variable, but with this implementation closing the slider does not.
This is not consistent with Fathom, but is consistent with handling of data interactives, in the sense that
deleting a data interactive leaves an orphan data context that cannot be reclaimed.