Fix bug: undo of delete selected cases.
Basically, the change causing parent keys to be validated broke this feature. Unfortunately the feature had not yet been QA'd. Otherwise it would have been recorded, properly, as a new bug.

The IChange object has an ambiguity in definition. Sometimes, (when the ultimate invoker is external), the parent property is a key. Other times it is the object itself. validateParent depended upon it being a key.

[#96680770]