Fix bug: Not Saved alert comes up after saving a new document.
Underlying problem was that asynchronous operations were not occuring in the expected sequence.
The document was declared dirty to cause a second save to finish the operation, but, with the removal of
autosave, this step was not automatically happening. Thus, the alert was coming up if you only saved one time.
[#100696640]

Also fixes [#100695988], "Original document has copy context file". The
copy operation was copying the current document's id as the parent,
regardless of whether we were creating an original file or a copy. Now,
the context document will not be affiliated to the wrong parent, but
having no affiliation, will not be displayed at all in the document
manger's file list, until a subsequent save occurs.