Cases get values from data items.
A lot of this was merged from the last failed branch. Up till now on this branch we were inserting into the data set, but it was a write only store. Now cases get their values from the underlying store and when we update a value we update the value in the case's whole extent.