Remove duplicate logging events
Each of these log events was logging an event that
was already being logged by the undo-redo controller
(via the `log` property of an undoable command).

The one exception is the rescaleScatterPlot event,
which previously logged both when triggered by a
user and when triggered by some other means. The
"by some other means" trigger will continue to log.

[#104009308]