Prevent initial comp opening from being undone
Previously, most of the components open during initialization
using the same undoable-command path as when a user opens
that component.

This allows us to distinguish opening a component during
initialization vs a user opening them. The `add[Component]`
functions that do not take an `isInitialization` argument
(such as the table or the calculator) already use a different
path for initialization, so this is not needed.

[#105875942]