Add ability to reduce multiple undoable commands
This adds the ability to take multiple undoable commands and merge
them into one single command that can be undone and redone.

This is implemented in an abstract way so that it can be used
flexibly in different scenarios. The simplest scenario, though,
is simply merging the `_beforeStorage` of the previous command,
and this can be implemented in just a  couple of lines.

Unlike the command-set pattern implemented in Building Models,
which adds an arbitrary number of commands to a single set, and
then undoes or redoes them all sequentially, this is more
suited to a situation where we might be streaming hundreds of
commands (e.g. when dragging a slider) and it would be ungainly
to want to store each action individually.

[#107317100]