Create new undoable action when color picker closed
With our previous implementation of merging color-change actions,
all consecutive color-change events on the same category would
be lumped together, even if the user closed the color picker
in between changes.

While this is fine, a more intuitive behavior seems to be that
closing the color picker should finalize one choice, and that
re-opening it should constitute a new action.

To do this, we update the PickerColorControl to optionally call
a new function, `closedFunc`, when the window is closed. (It
calls this with two arguments, the very initial color when the
picker was opened, and the final color, in case this is useful
to anyone, but these is not currently used by anything. Full
disclosure: those arguments were part of my initial implementation
of merging color-change events.)

When we compere consecutive commands for reducing, we add a random
number when the picker is first opened. When we close the picker,
we clear that number, so re-opening the picker will not reduce with
the previous commands.

[#107317100]