commitCurrentEdit on undo and redo
Previously, current editors remained in edit mode when undo
or redo was pressed, and their contents were not committed.

Note that, with the new text undoing and redoing using reducers
(in the next commit), this isn't actually necessary to get
good behavior from the text editor, but it seems to be good
practice to put `commitCurrentEdit` everywhere that we'd
expect a text field to have committed.

[#105880630]