Rewrite textEditor undo/redo using reducers
Previously, the textEditor depended on `commitCurrentEdit` in
order to create an undoable action.

This caused problems because (a) the document was being changed
without an undoable action being performed (which should now
be regarded as bad form), and (b) `commitCurrentEdit` wasn't
being reliably called. This was most noticeable when Undo was
pressed while the editor was still focused, but also
occasionally occurred at other random times.

The new pattern completely removes the dependence on
`commitCurrentEdit`. Each keystroke is an action that is
reduced with the previous keystroke.

We still use `commitCurrentEdit` to signal a break between
two actions (e.g. if the user is typing, clicks elsewhere,
and then resumes typing) but it is no longer a requirement
that `commitCurrentEdit` is called in order to correctly
create an undoable action.

[#105880630]