Fix sub-case ordering issues in build 0320.
The last fix changed the comparer. We were ordering case table groups by case id, but this is not going to be in the right order. So we fixed this to order by item index, but this was insufficient whent the hierarchies grew beyond two because we need to order groups within the parent groups and in turn within their parent groups. So, now to compare two groups we recursively search for a common ancestor and then take the order of their first children.