Initial work connecting CODAP to Cloud File Manager
This sets up a very simple global API under the DG namespace,
`DG.newDocument`, `DG.serializeDocument` and `DG.loadDocument`,
which is the majority of what is needed for the CFM (without
introducing forking yet).

The client connection only happens if CODAP is hosted inside
of an iframe with a CloudFileManager available in the parent.
This connects the CFM API to the new DG API.

The CFM example can be found by running the server and browsing
to http://localhost:4020/static/dg/en/current/resources/cfm/index.html

We will probably want to symlink this to a better location
when we deploy