Fixed sized horizontally-scrollable case tables.
We put the HierCaseTable contents in a scroll view and make the splitview render all its children at their natural sizes.
We make sure that a user cannot drag the table component wider than its contents and that, when the contents get smaller, they do not expose empty space on the right edge.

[#109649504]