Fix disappearing case table rows.
When you had scrolled down in a long case table and clicked on another component, it would often happen if you selected the Case Table component the scrolled table would appear blank until you scrolled in it again. This was found to be a consequence of component deselection and selection. During this process, the selected component is brought to the front by removing it from the DOM and adding it again. This loses scroll position.

We reset the scroll position to the value that Slickgrid thinks it has.

[#105266534]