Fix "case table remembers width of column for map documents"
This was due to DG.Component saving its content models in a singleton, DG.Component.contentMap.
We were not clearing this store with a new document.
This code has been in the codeline for a long time, but was not seen in the case table because
the case table did not have a persistent model.
 So far, it does not appear that this fix changes the behavior of other component types, but
 it is possible.

 Technical Debt: consider removing this and other global storage locations. At minimum these are potential
 memory leaks.

[#110090788]