Fix bug: Regrouped data context corrupted by Data Interactive on file restore.
If you drag attributes in the case table to create a new hierarchy and save the document, associated
data interactives would sometimes corrupt this organization when the document was restored.
This was because the flag blocking changes from the data interactive was not persisted with the document.

Note that the user may get a rather unfriendly alert on opening the document, saying that data changes are not
permitted.

[#108989676]