[#112545867] Add support for deleteCollection operation, fix questionable operation resets.
In a couple of instances we were editing the iChange.operation values to account
side effects. These have been cleaned up. In one case we submit an additional
operation to the queue with a completed status.

We add a position property to the object returned from getAttrRefByID to make it
match the object returned by getAttrRefByName.