SC 1.11 Upgrade: Remove call to setClass(..) in CaseTableView.render()
The call was appending to the class list, not resetting as in the prior equivalent API.
It looks as though the current version of sproutcore references the DOM element directly,
so the need for the call is obsolete. Keeping as commented out for now, in case it
proves that something (better) is needed.