[#98852476] Patch keyboard based selection
Disable shift/up-arrow and shift/down-arrow selection in the case
table row selection model until this behavior can integrate with CODAP selection
correctly. Till now keyboard shift and up or down arrows in the case table
would create the appearance of selection, but not inform CODAP, so selection
would not propagate. Making this behavior sensible requires a traditional
implementation of shift selection and ctl-key selection in the case tables and,
probably, the idea of navigation in the case table by keyboard control. For now,
we disable shift/arrow selection to avoid confusion.