[#115794901] Gracefully handle webViews that fail to load.
We put up a loading message in the background of the webview prior to load,
then when the load completes (perhaps from error) we change the message
to a load error. If the page load fails, especially because of a same origin error,
we cannot know about it, but the error page will be present.

If webView is a Data Interactive, and communicates with CODAP, we remove any
message from the background. This cleans things up for interactives with
transparent backgrounds.