Hide CFM menu options if `launchFromLara` url param
If CODAP is launched with `?launchFromLara=true`, this means that
it was launched from the document-server Launch Button that is
embedded in LARA, which means we should be a student.

When we launch as a student from LARA, we don't want the student to
have access to the CFM menus.

Currently we leave the blue menu bar, to support CODAP's help button
and any other custom buttons we may add, but this could be removed
if we decide it's not worth the extra space.

[#116548221]