New DI API: Move initiation of connection to data interactives to the Game Controller from the game view.
Arguably, this functionality belonged in the controller in the first place. It make the view cleaner and exclusively concerned with presentation.
Being so, it will be easier to integrate with DG.webView, if the opportunity presents itself.
In the mean time, this separation was necessary to centralize logic relating to whether the data interactive is connected and communicating,
and over what interface. This is needed for, among other reasons, whether to warn the user on deletion of the interactive.