[#117561145], [#121523735] Fix bugs: The "all changes saved" message no longer shows up. Can't download documents with large number of cases.
I found that the message was being set, but quickly wiped by a subsequent resetting of the dirty state.
Setting the dirty state to false was inadvertently setting the saved state to undefined.

This bug was present on current Chrome. Turned out that the 'download' dialog was using a data URI of the stringified JSON.
Appeared to bang up against URI size limits. Converted to a blob URL.