[#121640541] Provisional fix to table scrolling bug.
The underlying bug is a [bug in Chrome](https://bugs.chromium.org/p/chromium/issues/detail?id=466670).
It occurs only on retina (or retina-like) displays.
So, until this is fixed, we are forcing the scrollbar to be present. If normal scrolling freezes for scroll wheels,
the user will at least be able to use the scrollbar to navigate.