Close splash screen when document is loaded
This removes the time-based closing of the splash screen, and replaces
it with an explicit close in two locations: (1) showing the entry
dialog, and (2) opening a document.

Originally, even when a document was being opened directly via the
initial url, the entry dialog ("What would you like to do?") was being
shown. (This wasn't noticeable because it was covered by the splash
screen). We need to avoid this if we are going to open a document.
Therefore we look at the CFM and see if it has the properties indicating
it will open a document. (There may be a better way to do this.)