[#127416313] Fix bug: moving an attribute in the case table results in a scroll to the bottom.
The problem was we were using logic that was also used for case add, and this was blindly assuming
that new case was being added to the bottom and scrolling to it. We improve the case add logic to scroll
to the last added case, and do this only for case add.