[#131390457] Add ability to edit attribute characteristics.
We create a new editor for attribute properties. This is exposed
in the case table column header menus. It replaces the rename command.
It subtly changes the meaning of an attribute being editable.
If an attribute is not editable, the values can be changed,
but its properties can still be changed, including whether it
is editable.