[#132374669], [#131020987] Force CSV names to be unique and remove title line consideration
When loading CSV files force attribute names to be unique by appending numbers to the original name until one found unique.
To fix failing test for [#131020987], remove title line handling.
This was a feature that tried to detect if the first line of a CSV file was a title line by looking for first lines that had fewer columns than the table as a whole.
This was a problematic feature that was rarely used. We generally infer table name from file name, now.