Completely specify CFM/File menu in CODAP
Previously, a default menu was provided by CFM and CODAP customized it. In discussion with Jonathan we decided that CODAP should completely specify its menu configuration rather than share joint responsibility for it with CFM.

While there, fix the bug in which a document opened from or saved to Google Drive could not be saved locally [#134568775] by changing the behavior of the Save... menu item so that it always brings up the dialog (behaves like Save As...). In cases where autosave was enabled, the previous behavior was essentially a no-op, as the file would be saved every few seconds anyway. In situations where autosave was not enabled, the Save As... dialog would come up anyway. This way autosave handled saving to the current provider while the Save... menu item allows the user to change provider.