Added new query parameter (cfmBaseUrl) to load CFM from external url [#133897043]
If the cfmBaseUrl query parameter is present and it points to either codap.concord.org or a domain with no TLD like "localhost" or "dev" the globals.js and app.js files are appended to the url (after insuring a slash delimiter exists in the cfmBaseUrl).  If the cfmBaseUrl is not present or is not valid the static fallback urls are used.