Add FastClick.js library to fix bug which prevented column header menu from appearing - the hypothesis is that the menu was being opened via the touchstart event and then immediately closed via the "ghost" click event - aside from eliminating the 300ms delay between touchend and click events, the FastClick.js library also swallows the "ghost" click
Part of the confusing history if this bug is apparently explained by the fact that it was originally logged when FastClick.js wasn't part of CODAP. The FastClick.js library was added to CODAP along with jQueryUI as part of the original autocomplete implementation because it improved the behavior of the jQueryUI autocomplete library on iOS. This led to it appearing to be fixed. Then after the switch to CodeMirror for the formula editor, jQueryUI and FastClick.js were removed from CODAP, thus reintroducing the bug.