Fix file drop in Import Local File dialog
This code used to rely on the fact that both the CFM Local File Dialog drop handler and the CODAP document drop handler would both be triggered by a file drop on the dialog. This double-handling was eliminated in an earlier code cleanup on the CFM side (CFM now swallows the drop to prevent the double-dispatch) and so the workaround in CODAP code to prevent double-handling is no longer necessary.