Include React & ReactDOM in CODAP builds rather than relying on them being provided with CFM - use WebPack to combine React & ReactDOM into a single `codap-lib-bundle.js` - add `build:bundle-dev` and `build:bundle-prod` scripts to package.json - add `build`, `build:dev`, `build:production` scripts to package.json - `clean-and-build`, `clean-and-serve`, and `deploy:dev` scripts run bundle scripts as appropriate - `clean` script cleans the new `build` directory as well
Rather than continuing the pattern that CODAP has used up to this point of including all required libraries explicitly in the CODAP repository in the `libraries` and/or `resources` directories, the proposal here is that we specify the React dependencies in the `package.json` file and then require a build step to bundle them for use by CODAP. When starting from scratch with a new workspace, one would need to run the build step (`npm run build`) before continuing with normal development. Down the road, other dependencies could then be removed from the CODAP repository and bundled as well, e.g. CodeMirror, GeoJson-Utils, csv.js, es6-promise.js, Leaflet, etc., thus moving the dependency management from an implicit manual process to an explicit dependency expressed in the `package.json` and with management assistance from npm.