Fix Rerandomize Attribute functionality [#137053235]
While implementing the add case row in the case table recently, I came across the fact that DG.CaseTableDataManager.refresh() was notifying that the number of rows had changed on every change, whether or not the number of rows had actually changed. This clearly seemed to be incorrect, so I changed it to only notify when the number of rows changed, and when I didn't see any obvious indication that things were broken I left it in. It turns out Rerandomize Attribute was relying on that notification. Rather than using the onRowCountChanged notification inappropriately, however, we use the already defined onRowsChanged notification instead when the row count hasn't changed.