[#137311945] Make CODAP localizable in current sproutcore.
SC 1.11 apparently requires directories ending with .lproj instead of subdirectories of resources
directories to determine which languages to localize. E.g. english.lproj, spanish.lproj. Presumably
the language digraph would also work for the prefix. The idea is that the directory will have localized
resources.

So, we create these directories for english and spanish and move strings files there.