[#123902951] Provide full component creation support through DI API.
This is, among other things, in support of Embedded mode. The idea of
embedded mode is for a parent page to host a single instance of CODAP
as a layer above their page. In this mode there are no component
creation buttons. The hosting page will need to be able to completely
specify the position and contents of components.

In addition embedded mode requires the ability to retrieve information
about components, so better 'get' support is required.