[#114266699] Convert Euro Number format when importing semicolon-delimited file.
Non-English European numbers typically have comma as decimal separator and dot or blank as group separator.
CSV files are semicolon-delimited instead of comma delimited.
When importing these files we detect this and convert the european representation to US, then convert to number.
The CSV parsing library we were using was inadequate to the task so we employ a new one: [Papa Parse](http://papaparse.com).