Add attribute property 'hidden' to prevents table heading display.
[#138181403]

https://www.pivotaltracker.com/story/show/138181403

When a attribute defines a `hidden` property that is `true` it will not be displayed in the case tables.

When using the collaboration service (data interactive https://github.com/concord-consortium/codap-data-interactives/tree/master/CollaborationService )  The firebase internal ID's show up in the data tables.  This story is to add a new property to data attributes to mark an attribute as 'hidden'.