[#144560685] Fix bug: All web views maximized in standalone mode.
Modified the query parameter for standalone mode. It formerly took a value of 'true' or 'false'.
Now it can take the name of a plugin. If the name of a plugin is set, it will match the plugin
name with this value. This plugin and only this plugin will be sized to the full extent of the
screen. Other plugins will have the normal rules applied.