A new 'type' appears in the attribute properties dialog box for 'boundary'. Using this type, the name of the attribute need no longer conform to boundary naming conventions.
If the value of a boundary attribute begins with http or https, it will be treated as the url of a geojson object and made available to maps.

If the value of a boundary attribute is neither a geojson string nor a url, it will be treated as a state identifier and looked up in the state boundaries database.