Fix new attribute button tracking of component resize under some circumstances [#146957123]
The issue here was that the document container size wraps the content of the document and gets updated at the end of component resize operations, not during them. Rather than doing anything about that issue, we just don't iterate that far through the view hierarchy, as we only care about our immediate superviews for the purpose of placing the new attribute button.