[#147940119] Fix bug: When you edit the name of an attribute whose formula evaluates to a boundary, the thumbnail disappears in the case table.
We regularize value formatting in case table adapters. Some formatting
was occurring by getting a string value, some in the std cell formatter,
some in specialized formatter set on the column. Now all formatting, is
centralized to the cellformatter and on attribute values default type,
which may distribute problem to specialize formatters.