[#148791775] World boundaries do not come up in map
The particular data set had an attribute named 'Area' which triggered the map to regard it as the boundaries attribute. We're moving away from using special keywords for boundary attributes, and, to fix this bug we remove 'area' and 'Area' as reserved attribute names. We also now allow 'boundaries' and 'Boundaries'.