[#149889408] Fix bug: Can't add rows to an empty, non-flat table
This fix goes a bit beyond adding a proto-row to empty hierarchical
tables. When this was done, it was apparent that this removed an
affordance and felt confusing. So, now, we always have a proto-row.
When committing a proto-case for a child collection we create a
child case as child of last case in the parent collection.

Fix relationship line drawing to draw bottom line for last row.
It had been that the bottom line would sometimes not draw. We correct
this.

Fix an uncaught instance of an event outside run loop.