[#150486643] Improve how unexpected errors are handled.
These were errors that put up an unstyled error message saying a restart
is required. Effectively, they are hard crashes.

The error dialogs would only display in production and not in dev mode.

Found the errors occurred when an exception was uncaught within an SC.run
context. Found that in situations where the alert was not generated, the
exception caused the ending of the run-loop to be bypassed, meaning that
the instance is in an inaccurate state, at least until the next run loop.

* converted the fatal message to a dismissable alert.
* revised the SC exception handling to avoid corrupt state
* removed the differing behavior between production and dev modes.