[#150726368] Bug fix: Graph rescales when attribute is moved to parent collection
The problem was that during the brief interval in which the dragged attribute had no cases, the axis scale was set to the default with an animation. The animation was deferred, but meanwhile, the dragged attribute got added to the parent collection and acquired cases. The axis was then set to the correct bounds (the original bounds) but the animation to default bounds remained to be done, and wiped out the correct bounds.

Fix: Allow the plotAnimator to be reset, and do this in the situation in which there are bounds that need to be used.