[#151828321] CODAP crashes if an area map is present when new parent collection is created and then a case is selected.
The crash is caused by the MapPointLayer being notified of a change in case order and (in the base class) setting _mustMoveElementsToNewCoordinates to true. This flag, on selection, causes setCircleCoordinate to be called, and it has an assert that lat and long attributes exist. But the don't! Fix is to override notification to only set the flag when both are present.