[#157780336] Fix bug: Restore of set-aside cases "wonky"
In addition to cases reappearing in the wrong order, case identity seemed
impacted, so sorting in the case table would not work and selection was
inconsistent between case table and graphs. The cause was that the
algorithm updated CollectionClient.caseController array, but not
Collection.cases, so cases in a zombie state would get revived.

This checkin also includes improvements suggested in code review of
set-aside/delete cases performance improvement.