updates plots when data range has changed
rescaleOnParentCaseCompletion looks as if it was designed to rescale only if we add a new case, also appears to access cases incorrectly.  Add new function, rescalePlot to handle rescaling the plot when the range changes.  Consider further investigation of rescaleOnParentCaseCompletion to determine if the function is incorrectly written or if it needs to be removed.
[#125598121]