remove fast click
Remove fast click library from project which was used to reduce lag on touch screens.  The width=device-width in meta element sets the width of the page to follow the screen-width of the device (which will vary depending on the device).  With this tag, browsers assume you've made text readable on mobile, and the double-tap-to-zoom feature is dropped in favor of faster clicks.
[#159636714]