[#161420845] Fix bug CODAP fails to load in LARA.
This bug was mostly fixed for bug [#161299421], but this commit adds
extra protections.

The bug occurs because we used to create an Ur-document preceding any
document that would be visible to the user. The LARA CFM provider
apparently generates a log message during this period.

We test for the presence of the notification manager (handling notifications
to plugins) before calling it for log messages.