[#162347072] Bug fix: File with particular graph cannot be opened
A particular two-way dot chart was not coping well with minimized state on open. Now we hide the content view when a component is minimized to prevent such problems. For good measure we also don't attempt to draw a cell axis view if its frame has height or width of zero.