[#163795324] Bug Fix: Two legends get added to map instead of one
changeAttributeForLegend is a bottleneck for deciding whether an attribute can function as a legend for a graph or map. If the candidate attribute is a child of the childmost x, y, lat, long, or polygon attribute, it is _not_ should not become the legend attribute because for each point or polygon there will (typically) be multiple values.