[#163840831] Bug Fix: Graph axis tick labels become darker and fuzzier each time the graph is selected or unselected
Redrawing of the axes on component selection was put in place so that the "Click here" prompt could be drawn and removed. But it was added to _all_ axis views instead of just the plain ones where it was needed. The fix is to only observe if the given axis view is a DG.AxisView.