Fix bug which resulted in the generation of extraneous/redundant DG.CaseTableAdapters
In `updateTableAdapters()`, the code cleared the `caseTableAdapters` array before making calls to `findAdaptersForCollection()`, which tries to find previously created adapters in the now-cleared array. The result is extraneous/redundant DG.CaseTableAdapters, and since DG.CaseTableViews are generated for each adapter, extraneous/redundant DG.CaseTableViews were generated as well. Simply not clearing the array is sufficient to fix the problem.