Listen for CFM log messages and handle LARA data
This is designed to fix the issue where CODAP is embedded in SageModeler,
and never receives the `logLaraData` event in its LARA provider definition,
because the SageModeler LARA provider is being used instead.

The CFM is being updated to re-broadcast logLaraData events as a new
client event called "log." This will be received by any client that
uses `clientConnect`.

If CODAP receives a log event with the event type "logLaraData", it will
treat it the same as when "logLaraData" is called when it is at the
top level.

[#172071393]