[#176485373] Fix bug: Case table comes up in Component Mode with zero width.
The problem turned out to be that in Component Mode the Component is
immediately selected. This occurs before SlickGrid renders the contents,
so it is zero width. The view didn't take this possibility into account,
so sets the component width to match. We add this additional condition
to the readiness test.