[#185501866] Bug fix: Problem with plugins and global values
From Jonathan:
I found a small bug in CODAP in creating Globals from a document. It means that when a Global is created by loading a document it is not fully knowable to other Components. With the Sonification plugin, create graph feature (in progress, not yet in codeline), it causes an Error popup when creating a graph through the plugin. This is using the Document API, so the same behavior would probably exist if Globals are manipulated in the Story Builder plugin.
The fix is a one-liner. In the file, document_model.js, line 125, replace:
DG.GlobalValue.createGlobalValue(gv);
with
DG.globalsController.createGlobalValue(gv);